requirements.txt
livekit-agents==1.2.18
livekit-plugins-deepgram==1.2.18
livekit-plugins-elevenlabs==1.2.18
livekit-plugins-openai==1.2.18
livekit-plugins-silero==1.2.18
python-dotenv==1.2.1
